// apps/fleet/fleet-mfe/src/app/fleet-product/line/shared/enums/intent.enum.ts
export enum IntentEnum {
  POSTPAID = 'postpaid',
  PREPAID = 'prepaid',
  SMARTPHONE = 'smartphone',
  BASEPRO = 'basepro',
  PRODUCT_MULTIPLE = 'product-multiple',
  ADD = 'add',
  MODIFY = 'modify',
  PROSPECT = 'prospect',
  DEFAULT = 'default',
  PRODUCT_OPTIONS = 'options',
  MODEM_SWAP = 'modem-swap',
  INSTALLABLE_PRODUCT = 'installable-product',
  HARDWARE = 'hardware',
  MANUAL_ADDRESS = 'manual-address',
  MOVE = 'move',
  FIXED_IP = 'fixed-ip',
  FOURG_BACKUP_SERVICE = '4g-backup-service',
  CONTENT_PACK = 'content-pack',
  INTENT_DOMAIN_NAME = 'domain-name',
  STREAMING_SERVICE = 'streaming-service',
  BASE_FIXED_PRODUCT = 'base-fixed-product',
  FIBER_INSTALL = 'fiber-install',
  BASE_DROP_CABLE_MISSING = 'base-drop-cable-missing',
}