// apps/fleet/fleet-mfe/src/app/fleet-product/line/shared/enums/product-type.enum.ts
export enum ProductTypeEnum {
  MOBILE = 'MOBILE',
  OPTION = 'OPTION',
  SMARTPHONE = 'SMARTPHONE',
  SMARTPHONES = 'SMARTPHONES',
  INTERNET = 'INTERNET',
  DTV = 'DTV',
  TV = 'OTTV',
  BUNDLE = 'BUNDLE',
  TELEPHONE = 'TELEPHONE',
  MOBILE_LINE = 'MOBILE LINE',
  IDTV_LINE = 'IDTV LINE',
  TV_BOX = 'TV_BOX',
  HARDWARE = 'HARDWARE',
  EXPERIENCE = 'EXPERIENCE',
  DOMAIN_NAME = 'DOMAIN NAME',
}